backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Mail==0.9.1
pyodbc==5.3.0
python-dotenv==1.0.0
textblob==0.17.1
PyJWT==2.8.0
Werkzeug==3.0.1